refactor(mcp/ops): split server monolith into per-domain modules
Decompose mcp/ops/index.js (3,272 LOC) into a thin composition root
plus cohesive domain modules, keeping the public MCP tool surface,
request/response shapes, capability classes, and trust metadata
unchanged. tool-surface.test.js passes unmodified as the wire-contract
oracle.

- index.js: composition root only (server, policy load,
  registerAllOpsTools aggregation seam, signal handlers, entrypoint).
- Support modules: aws.js, db.js, respond.js, schemas.js, reconcile.js,
  github-actions.js.
- tools/: 12 per-domain registrars (logs, ec2, ecs, secrets, ssm,
  database, ranges, risk, images, s3, cost, approve), each registering
  via the existing registerTool with runtime deps injected for
  focused, MCP-free handler tests.
- Add reconcile.test.js: fake-client tests for the orphan
  classification helpers.
- Relocate the ADR-004-R14 state-bucket exception to tools/s3.js.

Part of #690
