+++ changelog.d/690.changed.md
@@ -1,0 +1,1 @@
+Internal refactor: split the `mcp/ops` server monolith (`index.js`) into a thin composition root plus per-domain tool modules under `mcp/ops/tools/` and support modules (`aws.js`, `db.js`, `schemas.js`, `reconcile.js`, `github-actions.js`, `respond.js`). Public MCP tool names, request/response shapes, capability classes, and trust metadata are unchanged.
