+++ docs/adr/exceptions.yaml
@@ -267,10 +267,10 @@
   {
     "rule_id": "ADR-004-R14",
     "owner": "@Brad-Edwards",
-    "reason": "Operational state-bucket map in the mcp/ops server, which reads live Terraform state from these UUID-suffixed infra/state buckets at runtime. The bucket names do not embed the AWS account ID. Moving them to runtime config is a separate change to the agent's working-day surface and out of scope for the identifier-strip; retained explicitly here.",
+    "reason": "Operational state-bucket map in the mcp/ops server, which reads live Terraform state from these UUID-suffixed infra/state buckets at runtime. The bucket names do not embed the AWS account ID. Relocated from index.js to the terraform_state tool module during the #690 modularization (still composed via registerAllOpsTools). Moving them to runtime config is a separate change to the agent's working-day surface and out of scope for the identifier-strip; retained explicitly here.",
     "expires_on": "2027-06-20",
     "paths": [
-      "mcp/ops/index.js"
+      "mcp/ops/tools/s3.js"
     ]
   },
   {
